+++ src/components/button/DraggableButton.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import { useDrag, DragSourceMonitor } from 'react-dnd';
-import { draggableButton } from './DraggableButton styles';
+import { draggableButton } from '../draggable/DraggableButton styles';
 import Button from '../button/button';
 import Icon from '../icon/icon';
 import { IconName } from '../icon/iconName';
 import { Colors } from '../../shared/constants/colors';
 
-import DraggableSourceProps from './DraggableSource.types';
+import DraggableSourceProps from '../draggable//DraggableSource.types';
 import ButtonProps from '../button/button.types';
+import useDragSpecs from '../draggable/useDragSpecs';
 
 type MergeProps = DraggableSourceProps & ButtonProps;
 
@@ -17,17 +17,10 @@ const DraggableButton: React.FC<MergeProps> = ({
     children,
     ...buttonProps
 }) => {
-    const specs = {
-        ...dragTargetConfiguration,
-        collect: (monitor: DragSourceMonitor) => {
-            onCollect(monitor);
-            return {
-                isDragging: monitor.isDragging(),
-            };
-        },
-    };
-
-    const [collectedProps, drag] = useDrag(specs);
+    const [collectedProps, drag] = useDragSpecs({
+        dragTargetConfiguration,
+        onCollect,
+    });
 
     return (
         <div ref={drag}>
