+++ src/components/button/index.ts
@@ -1,4 +1,5 @@
 import { ButtonType } from './buttonType';
+import DraggableButton from './DraggableButton';
 
 export { default } from './button';
-export { ButtonType };
+export { ButtonType, DraggableButton };
