+++ src/components/card/index.ts
@@ -1,1 +1,4 @@
+import DraggableCard from './DraggableCard';
+
 export { default } from './card';
+export { DraggableCard };
