+++ src/components/draggable/index.ts
@@ -1,6 +1,4 @@
-import DraggableButton from './DraggableButton';
 import DraggableTarget from './DraggableTarget';
-import DraggableCard from './DraggableCard';
 
 export { default } from './DraggableContainer';
-export { DraggableTarget, DraggableCard, DraggableButton };
+export { DraggableTarget };
